@@ -33,14 +33,14 @@
         "@testing-library/react": "^14.0.0",
         "@types/crypto-js": "^4.1.1",
         "@types/node": "^20.2.5",
-        "@types/react": "^18.0.37",
-        "@types/react-dom": "^18.0.11",
+        "@types/react": "^18.2.33",
+        "@types/react-dom": "^18.2.14",
         "@typescript-eslint/eslint-plugin": "^6.0.0",
         "@typescript-eslint/parser": "^6.0.0",
         "@vitejs/plugin-react": "^4.0.0",
         "@vitest/coverage-c8": "^0.33.0",
         "dotenv": "^16.1.4",
-        "eslint": "^8.44.0",
+        "eslint": "^8.52.0",
         "eslint-config-prettier": "^8.8.0",
         "eslint-plugin-import": "^2.27.5",
         "eslint-plugin-react-hooks": "^4.6.0",
