@@ -1,15 +1,16 @@
-import App from './App.tsx';
-import './index.css';
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
+import './index.css';
 import { DevSupport } from '@react-buddy/ide-toolbox';
 import { ComponentPreviews, useInitial } from './dev';
+import App from './App.tsx';
 
-ReactDOM.render(
-    <React.StrictMode>
-        <DevSupport ComponentPreviews={ComponentPreviews} useInitialHook={useInitial}>
-            <App />
-        </DevSupport>
-    </React.StrictMode>,
-    document.getElementById('root'),
+const element = document.getElementById('root') as HTMLElement;
+const root = ReactDOM.createRoot(element);
+root.render(
+  <React.StrictMode>
+    <DevSupport ComponentPreviews={ComponentPreviews} useInitialHook={useInitial}>
+      <App />
+    </DevSupport>
+  </React.StrictMode>,
 );
