@@ -1,15 +1,19 @@
 import Header from 'components/Header.tsx';
 import Footer from 'components/Footer.tsx';
+import Main from 'components/Main.tsx';
+import './App.css';
+import { BrowserRouter } from 'react-router-dom';
 
 function App() {
-    return (
-        <>
-            <Header />
-            <h3>Hello World</h3>
-            <p>extra</p>
-            <Footer />
-        </>
-    );
+  return (
+    <>
+      <BrowserRouter>
+        <Header />
+        <Main />
+        <Footer />
+      </BrowserRouter>
+    </>
+  );
 }
 
 export default App;
